+++ src/Scaffolder/Callbacks/TransformCallback.php
@@ -2,12 +2,14 @@
 
 namespace Phabalicious\Scaffolder\Callbacks;
 
+use Phabalicious\Exception\TransformFailedException;
 use Phabalicious\Method\TaskContextInterface;
 use Phabalicious\Scaffolder\Transformers\DataTransformerInterface;
 
 class TransformCallback implements CallbackInterface
 {
 
+    const TRANSFORMER_INPUT_FILENAME = 'transformer_input_file';
     protected static $transformers = [];
 
 
@@ -49,19 +51,16 @@ class TransformCallback implements CallbackInterface
      * @param $transformer_key
      * @param $files_key
      * @param $target_folder
+     * @throws TransformFailedException
      */
     protected function transform(TaskContextInterface $context, $transformer_key, $files_key, $target_folder)
     {
         $data = $context->get('scaffoldData');
         $tokens = $context->get('tokens');
 
-        $files = $data[$files_key] ?? [];
         /** @var DataTransformerInterface $transformer */
         $transformer = self::$transformers[$transformer_key] ?? false;
 
-        if (empty($files)) {
-            throw new \InvalidArgumentException('Could not find key in scaffold file ' . $files_key);
-        }
         if (!$transformer) {
             throw new \InvalidArgumentException(sprintf(
                 'Unknown transformer %s, available transformers %s,',
@@ -73,8 +72,14 @@ class TransformCallback implements CallbackInterface
         $target_path = $tokens['rootFolder'] . '/' . $target_folder;
 
         $context->io()->comment(sprintf('Transforming %s ...', $files_key));
-
-        $result = $transformer->transform($context, $files, $target_path);
+        
+        try {
+            $result = $transformer->transform($context, [$files_key], $target_path);
+        } catch (\Exception $e) {
+            if ($file = $context->getResult(self::TRANSFORMER_INPUT_FILENAME)) {
+                throw new TransformFailedException($file, $e);
+            }
+        }
 
         $context->io()->progressStart(count($result));
         foreach ($result as $file_name => $file_content) {
@@ -83,7 +88,9 @@ class TransformCallback implements CallbackInterface
             if (!is_dir($dir)) {
                 mkdir($dir, 0775, true);
             }
-            file_put_contents($full_path, $file_content);
+            if (false === file_put_contents($full_path, $file_content)) {
+                throw new \RuntimeException(sprintf("Could not write to file `%s`", $full_path));
+            }
             $context->io()->progressAdvance();
         }
         $context->io()->progressFinish();