+++ src/Scaffolder/Transformers/YamlTransformer.php
@@ -2,7 +2,9 @@
 
 namespace Phabalicious\Scaffolder\Transformers;
 
+use Phabalicious\Exception\TransformFailedException;
 use Phabalicious\Method\TaskContextInterface;
+use Phabalicious\Scaffolder\Callbacks\TransformCallback;
 use Symfony\Component\Yaml\Dumper;
 use Symfony\Component\Yaml\Yaml;
 
@@ -11,9 +13,11 @@ abstract class YamlTransformer implements DataTransformerInterface
 
     /**
      * Iterate over a bunch of yaml files.
+     *
      * @param TaskContextInterface $context
      * @param array $files
      * @return \Generator
+     * @throws TransformFailedException
      */
     protected function iterateOverFiles(TaskContextInterface $context, array $files)
     {
@@ -32,6 +36,10 @@ abstract class YamlTransformer implements DataTransformerInterface
                     yield $data;
                 }
             } elseif (in_array(pathinfo($filename, PATHINFO_EXTENSION), ['yml', 'yaml'])) {
+                $context->getConfigurationService()->getLogger()->debug(
+                    sprintf("Transforming file `%s` ...", $filename)
+                );
+                $context->setResult(TransformCallback::TRANSFORMER_INPUT_FILENAME, $filename);
                 $data = Yaml::parseFile($filename);
                 if ($data) {
                     yield $data;